First functional implementation
This is the first functional implementation of OpenData.
On file "participatorio_opendata.py" the connection to DB is opened and the functions that write the XML files are called. These function are "write_users_section" and "write_groups_section", that are defined on lib_user_section.py and lib_group_section.p files, respectively.

Signed-off-by: Bruno Nocera Zanette <brunonzanette@gmail.com>