Implement a Better solution to Comma's problem
This commit implemets a more elegant solution to the comma's problem.
On the function call it is given whether the flag is True (to use comma)
or False (to not comma), or an expression.
This expression is based on total rows and the current row.
If row is less then max row it is True, so it will use comma.

Signed-off-by: Bruno Nocera Zanette <brunonzanette@gmail.com>