Fix Bad Chars on Text (1st JSON Functional Release)
Fixed the problem of invalid chars in text, like', '\'. These chars crash the parser because the parser identifies them as the end of the tag, and then gets confused with the rest of the text.

The solution was to include a '\' in front of these chars on the small tags, like names and titles, and codify the whole text when it is too big, like descriptions and messages. This process makes the text unreadable for humans, but it doesnt loose any information when it is decoded.

Signed-off-by: Bruno Nocera Zanette <brunonzanette@gmail.com>