Modify Support functions into a Class definition
This commit modifies the support functions for each version (xml and json)
into a class definition. This change was made to simplify the code.
For example, there is no need to pass the database and output file
as arguments anymore.

Also, it was implemented a function to update the indentation every time a
tag is opened or closed, so that there is no need to specify the indentation
every time a write tag functions is called.

Signed-off-by: Bruno Nocera Zanette <brunonzanette@gmail.com>